增加瀑布流展示示例，增加下拉刷新和加载更多，完善demo功能

+++ XTNews/XTNewsURL.m
@@ -39,7 +39,7 @@
 
 + (NSString *)pageIdentifyWithPageNumber:(NSInteger)aPageNumber
 {
-    return [NSString stringWithFormat:@"%ld-20",aPageNumber * 20];
+    return [NSString stringWithFormat:@"%ld-20",(long)aPageNumber * 20];
 }
 
 @end
